Added Permission Hours to Timesheet, version change, link to manage request in dashboard

+++ NLTD.EmployeePortal.LMS.Common/DisplayModel/TimeSheetConsolidate.cs
@@ -7,7 +7,7 @@ namespace NLTD.EmployeePortal.LMS.Common.DisplayModel
         public string name { get; set; }
         public TimeSpan hours { get; set; }
         
-        public int permissionCount { get; set; }
+        public decimal permissionCount { get; set; }
         public decimal leaveCount { get; set; }
         public int lateCount { get; set; }
         public int earlyCount { get; set; }
